integrate importing/exporting code in TrackingModel and gui in dockwidget
